skills/wechat-publisher/requirements.txt
openai>=1.0.0
httpx>=0.24.0
markdown>=3.4.0
Pillow>=10.0.0
